Initial baseline image (Ansible)
Create a baseline image and helper functions that can be used and
shared across provision scripts. Helper functions stored in lflib.sh
will be copied to /tmp/lflib.sh which can be sourced by other
provisioning scripts and reused.

Another change to the formatting is within the baseline script itself
to make it more logical to read and easier for multi-distro install
scripts to maintain. We will install packages organized by function
with each function resposible for installing a single logical unit
of software packages so that package install scripts are closer
together.

Issue: RELENG-288
Change-Id: I96d1ec83cec82130d30a245c39b6ac8018f92428
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>