Revert "Bootstrap Ansible if not available"
This is not actually necessary. We need Ansible in the VM that's running the test not the VM being built. Which means the VM attached in to Jenkins needs to have Ansible installed in the first place. We just need to ensure that Jenkins is initially configured with one (which we currently have to do anyway even for the regular shell script provisioner).

This reverts commit 99f875674576f3643c9713750a7017f581ff758b.

Change-Id: I607526f4f6575e2a29c7184e0196e9c92040afd9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>