Update patch set 2
Patch Set 2:

> (2 comments)
 > 
 > Previously with RS, we never used ssh_proxy configuration while
 > building packer images etc..., has this been put in place due to
 > any limitation with the cloud provider?

The ssh_proxy thing is new with our current configuration for CI cloud deployments. If you've noticed, there is _no_ way for you to access any minions directly from your machine even if you're on the VPN. That's because the minion network does not have routing to or from the core LF networks. The only routing that they have is to the pet cloud (a separate tenant) and to the internet.

Patch-set: 2