Expand rendered image name to include sec and ms
Expand the rendered image name to include seconds and milli-seconds.
This reduces the probability of image name clashes in the env.

The initall change Ic1853ba7566def397b8bce389e2779fe3020c7ce was
reverted since this had issues bringing up an instance while
building the image on the OpenStack cloud provider because the
the default instance_name is taken as the image_name, when one is
not provided. The instance_name does not allow FQDN format
hostname ending with a period+number.

The workaround is to use separate `instance_name` paramter
in your builder which does not end with period+number.

Tested the change to work:
"ZZCI - CentOS 7 - docker - 20180605-015801.089"

JIRA: RELENG-971
Change-Id: Ia195bd99d55f23b4f39acf590520637e54fc006a
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>