Expand rendered image name to include sec and ms
Expand rendered image name to include seconds and milliseconds.
This reduces the probability of image name clashes in the env.

The initial change Ic1853ba7566def397b8bce389e2779fe3020c7ce was
reverted, since the change had issues bringing up an instance on the
OpenStack cloud provider. This was because the default
`instance_name` value is taken as the `image_name`, when the
instance_name is unset, the instance_name does not allow FQDN format
hostname ending with a period+number.

The workaround is to use separate `instance_name` parameter in your
builder which does not end with period+number.

Tested the change to work:
"ZZCI - CentOS 7 - docker - 20180605-015801.089"

JIRA: RELENG-971
Change-Id: Ia195bd99d55f23b4f39acf590520637e54fc006a
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>