Update patch set 1
Patch Set 1:

> Is there any benefit of using the uuid in the `instance_name`,
 > making it less human readable? I prefer not having the uuid since
 > builder param represents a "name" which should be something
 > understandable.

Yes, we are not stuck with a 1 second level granularity. I has a much lesser chance of instance_name collision as uuid is significantly more random then timestamp.

The instance_name in the definition is already prefixed with centos7-builder-###### we don't need to care about what the trailing hash is.

Patch-set: 1