Update patch set 1
Patch Set 1:

> > Is there any benefit of using the uuid in the `instance_name`,
 > > making it less human readable? I prefer not having the uuid since
 > > builder param represents a "name" which should be something
 > > understandable.
 > 
 > Yes, we are not stuck with a 1 second level granularity. I has a
 > much lesser chance of instance_name collision as uuid is
 > significantly more random then timestamp.
 > 
 > The instance_name in the definition is already prefixed with
 > centos7-builder-###### we don't need to care about what the
 > trailing hash is.

I think using a UUID in the instance_name is probably smart. As Thanh points out, we're prefixed with ${distro}-${template}- already so it will be more identifiable. The image_name is in general what I care more about as being fully human readable anyway.

My alternative suggestion was going to be just append ' - packer' to then dof the image_name with everything else we've got. The only reason for that would be if we shift build image tools in the future we would know what tool built an image, though really only helpful during a cutover to a new tool.

Patch-set: 1
Label: Verified=0