Update patch set 1
Patch Set 1:

> Just to give some clarity here, the instance_name has nothing to do
 > with the minion image name (centos7-builder-######), it's fleeting
 > and alternative to (image_name value) to spawn up a node to build
 > the packer image for that template.

Yes that's right. It's the name of the node instance. If it were to be managed by Jenkins openstack plugin that would be the "prd-centos7-builder-2c-2g-#####" node name. Which OpenStack plugin used to use a random hash for, now they simplified it to use an incremental number instead.

What we're saying here is we don't care about the instance_name so it doesn't need to be human readable. Setting a hash value here will give us better protection against 2 server instances spawning at the same time with the same name.

Patch-set: 1