Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > Hmm yeah I think that will work.  How would we know to pick the
 > correct image though?

Not following the statement / question. If you mean, how would you figure out which image to add / configure in Jenkins, then the variable that Trevor is suggesting would list the arch...

Overriding the variable to be something like arm64 would be done from the settings file specific to the image build.

So something like centos-7-arm64.json instead of centos-7.json (which would use the default of x86_64

Selecting which packer to then build is a matter of the job definitions just like it is now

Patch-set: 1
Label: Verified=0