Update patch set 1
Patch Set 1: Code-Review-2

This works but we should not merge this yet. We need to deploy this in phases. This depends on global-jjb jobs first being updated to use clouds.yaml for packer builds. Once that's done and we've deployed it to all clouds we can push this out after.

Patch-set: 1
Label: Code-Review=-2