Enable selection of instance knobs
We now operate in multiple Openstack clouds (even if from the same
provider) and thus need to properly enable selectability of the region,
availability zone, instance flavor size, and if a volume needs to be
created (and at what size) for builds. We default to our current
standard configuration for Vexxhost Montreal.

If vm_use_block_storage is set true then vm_volume_size will be used
(along with the image_disk_format option which must be "raw"). If it is
false (default) then the volume size and format will be ignored. This is
needed for building on flavors that do not predefine any attached disk
as is the case for all flavors in the SJC cloud for Vexxhost.

Change-Id: I358c414adb49bba9349f30a1190c2f8b90f13d14
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>