Remove unknown config value for aws template
Change c0dc5f030e259b70533a4d626eae5ab176fdd560 added in an
instance_name variable to the aws template. Unfortunately the
verification which happens in a cloud not using aws as the cloud itself
apparently couldn't pick up the fact that it's not valid with an EC2
definition. Trying to upgrade the LF internal Jenkins to latest
common-packer uncovered this problem.

Change-Id: Id96671cbe674c2af0df7ad4f4babed2b1ee8a4ed
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>