Update patch set 2
Patch Set 2:

I've done some recent work in this area. It's not necessarily Python 2 that's needed but Ansible needs Python to be available on the system. On Ubuntu 16.04 I think it didn't come with python by default so we needed something to ensure python-minimal got pulled in. In 18.04 and newer the script is no longer needed because python-minimal is provided by default.

If however we want to support Docker container images we will still need a similar script to ensure both sudo and python-minimal gets pulled in. This is what I'm using at the moment:

 #!/bin/bash
 # Docker images do not have sudo and python-minimal by default
 if ! command -v sudo; then
     if command -v apt-get; then
         apt-get update -y
         apt-get install -y sudo python-minimal
     fi
 fi

The assumption here is that if sudo is not available then packer's likely building a docker container because they are typically pretty minimal. In which case I assume we want to install sudo and python-minimal which are both minimum reqs for Ansible to work.

Patch-set: 2
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0