Change IUS configuration
While trying to troubleshoot why a recent CentOS7 system was failing to
get past the installation of yum-plugin-replace from IUS it was
discovered that IUS has changed up the keys that they are signing with.

Along with this change are changes to their official distribution point
according to the standard release package.

While we don't actually install the release package, we do have reason
to configure the rest of the repository correctly, as such we need to
follow the changes that were made back in May of 2019!

Change-Id: Ibace4cdf89d0615a2e5117f9ea3ca50730d5c93d
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>