Remove coala from common-packer
Move away from coala since that project refuses to produce any
new release and the version we're using is quite dated at this
point.

Coala does not seem to be working well with prettier pre-commit
hook in place, since prettier is auto-formating the yaml code
to make the code terse which fails the tox verify job.

Issue: RELENG-2641
Change-Id: I8d782822ca87add9108a044fd6e19b1c23686717
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>