Update patch set 10
Patch Set 10:

Sorry to piggyback on this, but I need to run a test with this change. But my question is: do we NEED the older version for 16.04? Because if not, we can just check for the arch. Even the 0.3.ds1-28+b1 is almost 3 years old, so I'm hoping it is compatible with both. The reason I'm hoping this works both ways is that there's no arm64 version of 0.3.ds1-26, so it would not work for 16.04+arm64 builds.

Patch-set: 10
Reviewer: Gerrit User 192 <192@f2a2391d-0403-4fa5-b183-4056999a8a09>