Use official CentOS & Ubuntu images via prodId
The previous filter method could result in getting an unexpected image
even one that's entirely different than what is intended. This patch
uses a product-code search of the marketplace for the official upstream
supported images rather than the custom ones provided by Amazon.

This patch is based on Zack William's suggestion on the lf-releng
mailinglist.

Reference:
https://lists.linuxfoundation.org/pipermail/lf-releng/2020-February/000135.html

Change-Id: Ia6208bd7a61036eb210d07533e3ceedbbc787678
Co-authored-by: Zack Williams <zdw@opennetworking.org>
Signed-off-by: Thanh Ha <zxiiro@gmail.com>