Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > Running "git --version" on the built system returns this:
> > git version 1.8.3.1
> > 
> > If you look at the build log, it seems that the "Replace git with git224" step is skipped: https://jenkins.opendaylight.org/sandbox/view/All/job/builder-packer-verify-build-centos-7-builder/1/console
> 
> Hmm, odd it is not skipped when run with Docker I wonder what's tripping the when condition when run in Vexxhost. I will investigate.

Ok it turns out I didn't test the "when" condition in my Docker based setup so that was why I thought it worked.

Solution is to class cast the ansible_distribution_major_version to int as that variable actually returns a AnsibleUnsafeText datatype which cannot be compared to an int value.

Retried the build (https://jenkins.opendaylight.org/sandbox/view/All/job/builder-packer-verify-build-centos-7-builder/2/console) and it appears to have not skipped installing git224 this time. I'm not sure if I have the power to start the VM that is produced from this job though.

Eric, when you get a chance can you spin up the VM from this job and double check the Git version?

Patch-set: 2