Add support for Ubuntu 20.04 and newer
Ubuntu 20.04 no longer has python-* packages and instead the packages
are now more explicitly either python2-* or python3-* with packages
named python-is-python[23] to set the default. This patch when run
on Ubuntu systems will check if python-is-python3 is available and
installs that as the system python version, otherwise installs
python-minimal.

Also adjust the comments since they are no longer accurate or in the
right places.

Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: I830e0250ddbee395fdeddd78fcc7824f7855d75d