Rename vexxhost to openstack for builder names
Considering the AWS builders are named AWS after the service let's
rename the vexxhost ones to openstack. As far as I am aware we are
not supporting multiple providers with the same backend at the same
time so this will make things consistent.

When we later add docker support it will just be called "docker" as
well.

Issue: RELENG-1052
Change-Id: I8f46d544961e508cdeee28d4e761c2feae5e744e
Signed-off-by: Thanh Ha <zxiiro@gmail.com>