Remove pre-commit from tox envlist
Pre-commit is causing errors throughout CI. This appears to be based
on upstream libraries that we do not control, so even pinning
versions will not fix the issue. Until these upstream problems have
been fixed, the best solution is to fully pull the pre-commit env
from the default list.

Issue: RELENG-3344
Change-Id: I07daf63bcb03252367b11a64f44c938bf807eea6
Signed-off-by: Eric Ball <eball@linuxfoundation.org>