Set packer builder type with env variable
Using the -only CLI option for packer build does not work as
expected. The #CR 90dd675df482a8 (global-jjb) submitted with
common-packer v0.7.0 fails to build packer jobs because of a
bug in the packer build CLI.

The `-only` option packer build CLI does not set the cloud
provider name/type as per the documentation. The alternative
is to use an environment variable that is passed through the
packer job and can be reference in the packer templates.

Issue: https://github.com/hashicorp/packer/issues/10495
Change-Id: I1c1e2e7c5545df5a2a374a02714634a0f4086378
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>