Use workaround for -only option
Using the -only option to set the cloud provider
type/name does not work as documented.

A workaround can be used to pass the env variable to
set the cloud provider name/type. The $PACKER_BUILDER
environment variable is set in global-jjb which is set
to the packer jobs.

Issue: https://github.com/hashicorp/packer/issues/10495
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>
Change-Id: I1c1e2e7c5545df5a2a374a02714634a0f4086378