Feat: Add Openstack devstack template
Move the Openstack devstack templates to build VM's with devstack
installed and ready for testing. ORAN community has requested devstack
images to be available on the CI env.

The templates and provision files have been tested to build the latest
version of the devstack yoga on the CI.

ZZCI - CentOS Stream 8 - devstack-yoga - x86_64 - 20230612-132235.763

Change-Id: I60d37d769158f34b22b84111fd5229c616232c27
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>