Update install-python.sh for Ubuntu 24.04+
Running install-python.sh on Ubuntu 24.04 fails during the netselect/
select_fastest step. This can be safely skipped on 24.04, but is still
useful up to 22.04, so we check lsb-release for a DISTRIB_RELEASE
major version less than 24.04 before entering the step.

This also includes a small fix to indentation in the case block (lines
were previously indented to 3 spaces instead of 4).

Change-Id: I4ffb398d78abcd3cf1bd1ec14a034b21d66ebab3
Signed-off-by: Eric Ball <eball@linuxfoundation.org>